vérification de la version au démarrage (suite) - limitation à une verif auto par jour - possibilité de désactiver les vérifs autos paramètres proxy dans ParametreDlg paramètre timeout dans ParametreDlg (utilisé dans curl_util --> uniquement par la verification de version) amélioration des performances à la sauvegarde des paramètres (ParamManager::Set) suppression de divers warnings de compilation
git-svn-id: https://svn.kenai.com/svn/livrotheque~livrotheque-repository@38 42eda3aa-e831-e092-880c-b46d1eaaeeec

+++ src/livre_html.cpp
@@ -267,7 +267,7 @@ size_t LIVRE_HTML::write_d(void *ptr,size_t size, size_t nmemb, void *stream)
  
 wxString LIVRE_HTML::recup_donnee(wxString donnee, wxString debut, wxString fin, int type_recup) {
     wxString resultat;
-    size_t debut_d, fin_d;
+    int debut_d, fin_d;
     //int fin_d;
     
     debut_d=donnee.Find(debut);
@@ -286,7 +286,7 @@ wxString LIVRE_HTML::recup_donnee(wxString donnee, wxString debut, wxString fin,
     resultat.Replace("<br />","\n",true);
     resultat.Replace("<BR />","\n",true);
     // suppression des blancs de début et de fin
-    int i=0;
+    size_t i=0;
     while ((resultat[i]==' ' || resultat[i]=='\t' || resultat[i]=='\n') && i<resultat.Len()) {
         i++;
     }
